# Optimize dataset building with precomputed indices