## Summary
- precompute node column indices and pump control arrays in dataset builders
- replace repeated DataFrame lookups with direct numpy indexing

## Testing
- `pytest`
- timed `build_dataset` and `build_sequence_dataset` before/after (7.46s→2.11s, 1.38s→0.47s)


------
https://chatgpt.com/codex/tasks/task_e_688fa27c07f88324b25ea68a0174b63a